Seperate ConsequenceDashboard to a independent file
