+++ factory-ai-vision/EdgeSolution/modules/WebModule/ui/src/components/LiveViewDashboard/LiveViewDashboard.tsx
@@ -1,4 +1,4 @@
-import React, { useEffect, FC, useState } from 'react';
+import React, { useEffect, useState } from 'react';
 import { Flex, Text, Loader, Alert } from '@fluentui/react-northstar';
 import { useDispatch, useSelector } from 'react-redux';
 
@@ -9,11 +9,12 @@ import {
   thunkGetInferenceMetrics,
   resetStatus,
 } from '../../store/project/projectActions';
-import { Project, Status as CameraConfigStatus, TrainingMetrics } from '../../store/project/projectTypes';
+import { Project, Status as CameraConfigStatus } from '../../store/project/projectTypes';
 import { State } from '../../store/State';
 import { LiveViewContainer } from '../LiveViewContainer';
 import { InferenceMetricDashboard } from './InferenceMetricDashboard';
 import { Button } from '../Button';
+import { ConsequenceDashboard } from './ConsequenceDashboard';
 
 export const LiveViewDashboard: React.FC<{ projectId: number }> = ({ projectId }) => {
   const { error, trainingLog, status, trainingMetrics } = useSelector<State, Project>(
@@ -93,48 +94,4 @@ const useAllTrainingLog = (trainingLog: string): string => {
     setAllLogs((prev) => `${prev}\n${trainingLog}`);
   }, [trainingLog]);
   return allLogs;
-};
-
-interface ConsequenceDashboardProps {
-  trainingMetrics: TrainingMetrics;
-}
-const ConsequenceDashboard: FC<ConsequenceDashboardProps> = ({
-  trainingMetrics: { curConsequence, prevConsequence },
-}) => {
-  return (
-    <table style={{ textAlign: 'center', width: '60%' }}>
-      <tr>
-        <td style={{ width: '200px' }}></td>
-        <td>Precision</td>
-        <td>Recall</td>
-        <td>mAP</td>
-      </tr>
-      <tr>
-        <td>Updated Model Metrics</td>
-        <td style={{ color: '#9a0089' }}>
-          {curConsequence?.precision === null ? '' : `${((curConsequence?.precision * 1000) | 0) / 10}%`}
-        </td>
-        <td style={{ color: '#0063b1' }}>
-          {curConsequence?.recall === null ? '' : `${((curConsequence?.recall * 1000) | 0) / 10}%`}
-        </td>
-        <td style={{ color: '#69c138' }}>
-          {curConsequence?.mAP === null ? '' : `${((curConsequence?.mAP * 1000) | 0) / 10}%`}
-        </td>
-      </tr>
-      {prevConsequence && (
-        <tr>
-          <td>Previous Model Metrics</td>
-          <td style={{ color: '#9a0089' }}>
-            {prevConsequence?.precision === null ? '' : `${((prevConsequence?.precision * 1000) | 0) / 10}%`}
-          </td>
-          <td style={{ color: '#0063b1' }}>
-            {prevConsequence?.recall === null ? '' : `${((prevConsequence?.recall * 1000) | 0) / 10}%`}
-          </td>
-          <td style={{ color: '#69c138' }}>
-            {prevConsequence?.mAP === null ? '' : `${((prevConsequence?.mAP * 1000) | 0) / 10}%`}
-          </td>
-        </tr>
-      )}
-    </table>
-  );
 };
